+++ UIKit/SpecHelper/Extensions/UICollectionViewCell+Spec.h
@@ -4,4 +4,9 @@
 
 - (void)tap;
 
++ (instancetype)instantiatePrototypeCellFromStoryboard:(UIStoryboard *)storyboard
+                              viewControllerIdentifier:(NSString *)viewControllerIdentifier
+                                 collectionViewKeyPath:(NSString *)collectionViewKeyPath
+                                        cellIdentifier:(NSString *)cellIdentifier;
+
 @end