+++ UIKit/SpecHelper/Extensions/UITableViewCell+Spec.h
@@ -6,4 +6,9 @@
 - (void)tapDeleteAccessory;
 - (void)tapDeleteConfirmation;
 
++ (instancetype)instantiatePrototypeCellFromStoryboard:(UIStoryboard *)storyboard
+                              viewControllerIdentifier:(NSString *)viewControllerIdentifier
+                                      tableViewKeyPath:(NSString *)tableViewKeyPath
+                                        cellIdentifier:(NSString *)cellIdentifier;
+
 @end