+++ UIKit/SpecHelper/Extensions/UIView+Spec.h
@@ -2,6 +2,7 @@
 
 @interface UIView (Spec)
 - (UIView *)subviewWithAccessibilityIdentifier:(NSString *)accessibilityIdentifier;
+- (UIView *)firstSubviewOfClass:(Class)aClass;
 
 - (void)tap;
 - (void)swipe;